+++ app/src/main/java/com/itachi1706/busarrivalsg/AsyncTasks/PopulateListWithCurrentLocationRecycler.java
@@ -16,8 +16,8 @@ import com.google.gson.reflect.TypeToken;
 import com.itachi1706.appupdater.Util.ValidationHelper;
 import com.itachi1706.busarrivalsg.Database.BusStopsDB;
 import com.itachi1706.busarrivalsg.Fragments.BusStopNearbyFragment;
-import com.itachi1706.busarrivalsg.GsonObjects.Distance;
-import com.itachi1706.busarrivalsg.GsonObjects.LTA.BusStopJSON;
+import com.itachi1706.busarrivalsg.gsonObjects.Distance;
+import com.itachi1706.busarrivalsg.gsonObjects.sgLTA.BusStopJSON;
 import com.itachi1706.busarrivalsg.R;
 import com.itachi1706.busarrivalsg.RecyclerViews.BusStopRecyclerAdapter;
 import com.itachi1706.busarrivalsg.Util.StaticVariables;
