+++ app/src/main/java/com/itachi1706/busarrivalsg/Services/BusStorage.java
@@ -3,8 +3,8 @@ package com.itachi1706.busarrivalsg.Services;
 import android.content.SharedPreferences;
 
 import com.google.gson.Gson;
-import com.itachi1706.busarrivalsg.GsonObjects.Offline.BusArrayJSON;
-import com.itachi1706.busarrivalsg.GsonObjects.Offline.BusJSON;
+import com.itachi1706.busarrivalsg.gsonObjects.offline.BusArrayJSON;
+import com.itachi1706.busarrivalsg.gsonObjects.offline.BusJSON;
 import com.itachi1706.busarrivalsg.Objects.BusServices;
 
 import org.json.JSONArray;
