+++ app/src/main/java/com/itachi1706/busarrivalsg/gsonObjects/offline/BusArrayJSON.kt
@@ -1,4 +1,4 @@
-package com.itachi1706.busarrivalsg.GsonObjects.Offline
+package com.itachi1706.busarrivalsg.gsonObjects.offline
 
 /**
  * Created by Kenneth on 17/6/2015
